solve oveall dependency on rdkit
